refactor(web-design): consolidate assistant UI adapters under elements
