+++ web/design/scripts/agent-plan-contract-check.ts
@@ -1,5 +1,5 @@
 import { strict as assert } from "node:assert"
-import { fleetAgentPlanPresentation } from "../src/components/assistant-ui/fleet-agent-plan"
+import { fleetAgentPlanPresentation } from "../src/components/elements/fleet-agent-plan"
 
 const pending = [
   { id: "1", title: "Inspect the request", status: "pending" as const },
