+++ web/design/src/components/fleet-pi/chat/beui-tool-renderer.tsx
@@ -7,7 +7,7 @@ import { Citations } from "../../agents/citations"
 import { FileDiff } from "../../agents/file-diff"
 import { ImageGeneration } from "../../agents/image-generation"
 import { TodoList, type TodoItem } from "../../agents/todo-list"
-import { FleetAgentPlan, fleetAgentPlanPresentation, type FleetPlanItem } from "../../assistant-ui/fleet-agent-plan"
+import { FleetAgentPlan, fleetAgentPlanPresentation, type FleetPlanItem } from "../../elements/fleet-agent-plan"
 import { ToolApproval } from "../../agents/tool-approval"
 import { ToolResult, ToolResultOutput, type ToolResultStatus } from "../../agents/tool-result"
 import type { ToolRendererProps } from "../../agent-elements/tools/tool-renderer"
