+++ web/design/src/components/fleet-pi/chat/fleet-pi-agent-chat.tsx
@@ -17,7 +17,7 @@ import { cn } from "../../../lib/utils"
 import { GenerativeTextRenderer } from "../../openui/inline-renderer"
 import { PI_TOOL_RENDERERS } from "../pi/tool-renderers"
 import { FleetPiToolRenderer } from "./fleet-pi-tool-renderer"
-import { FleetReasoningPanel } from "../../assistant-ui/fleet-reasoning-panel"
+import { FleetReasoningPanel } from "../../elements/fleet-reasoning-panel"
 import { FleetTurnStatus } from "./fleet-turn-status"
 import { ToolTimeline, type TimelineStep } from "../../elements/tool-timeline"
 import { FilePenLine, FileSearch, Search, Terminal, Wrench, type LucideIcon } from "lucide-react"
